Fix after migration to Zoomdata 2.6.3
There was a timing issue, the query was returning quickly and the code was catching the event prior to saving the data to state.

This is a quick fix, need to look at rewriting the saga to use eventChannel to handle the thread events, which will require react/redux upgrade.